Change dashboard to populate center region and left region.
Currently views are responsible for creation of left and center regions but I'm not sure if this design is any good. Rationale behind that design is that left region contains tools / operations related to what is presented in main container so that view creates something in center and also tools on the left.

Actually what is needed is some design which allows to embed views as tabs of tab container in center region. Need to discuss that with someone.